OpenFeature: Share http.Transport across proxy requests
