+++ pkg/registry/apis/ofrep/proxy.go
@@ -3,14 +3,11 @@ package ofrep
 import (
 	"bytes"
 	"context"
-	"crypto/tls"
-	"crypto/x509"
 	"encoding/json"
 	"fmt"
 	"io"
 	"net/http"
 	"net/http/httputil"
-	"os"
 	"path"
 	"strconv"
 
@@ -102,15 +99,6 @@ func (b *APIBuilder) newProxy(proxyPath string) (*httputil.ReverseProxy, error) 
 		return nil, fmt.Errorf("OpenFeatureService provider URL is not set")
 	}
 
-	var caRoot *x509.CertPool
-	if b.caFile != "" {
-		var err error
-		caRoot, err = getCARoot(b.caFile)
-		if err != nil {
-			return nil, err
-		}
-	}
-
 	director := func(req *http.Request) {
 		req.URL.Scheme = b.url.Scheme
 		req.URL.Host = b.url.Host
@@ -118,23 +106,6 @@ func (b *APIBuilder) newProxy(proxyPath string) (*httputil.ReverseProxy, error) 
 	}
 
 	proxy := proxyutil.NewReverseProxy(b.logger, director)
-	proxy.Transport = &http.Transport{
-		TLSClientConfig: &tls.Config{
-			InsecureSkipVerify: b.insecure,
-			RootCAs:            caRoot,
-		},
-	}
+	proxy.Transport = b.transport
 	return proxy, nil
-}
-
-func getCARoot(caFile string) (*x509.CertPool, error) {
-	// It should be safe to ignore since caFile is passed as --internal.root-ca-file flag of apiserver
-	// nolint:gosec
-	caCert, err := os.ReadFile(caFile)
-	if err != nil {
-		return nil, err
-	}
-	caCertPool := x509.NewCertPool()
-	caCertPool.AppendCertsFromPEM(caCert)
-	return caCertPool, nil
 }
